Fix: Pb ajout suppression des boxes

+++ htdocs/admin/external_rss.php
@@ -1,7 +1,7 @@
 <?php
 /* Copyright (C) 2003      Éric Seigne          <erics@rycks.com>
  * Copyright (C) 2003,2005 Rodolphe Quiedeville <rodolphe@quiedeville.org>
- * Copyright (C) 2004-2006 Laurent Destailleur  <eldy@users.sourceforge.net>
+ * Copyright (C) 2004-2007 Laurent Destailleur  <eldy@users.sourceforge.net>
  * Copyright (C) 2004      Sebastien Di Cintio  <sdicintio@ressource-toi.org>
  * Copyright (C) 2004      Benoit Mortier       <benoit.mortier@opensides.be>
  *
@@ -39,15 +39,20 @@ $langs->load("admin");
 if (!$user->admin) accessforbidden();
 
 $def = array();
+$lastexternalrss=0;
 
 // positionne la variable pour le nombre de rss externes
-$result=$db->query("select count(*) nb from ".MAIN_DB_PREFIX."const WHERE name like 'EXTERNAL_RSS_URLRSS_%'");
+$sql ="select MAX(name) name from ".MAIN_DB_PREFIX."const";
+$sql.=" WHERE name like 'EXTERNAL_RSS_URLRSS_%'";
+$result=$db->query($sql);
 if ($result)
 {
     $obj = $db->fetch_object($result);
-    $nbexternalrss = $obj->nb;
+    eregi('([0-9]+)$',$obj->name,$reg);
+	if ($reg[1]) $lastexternalrss = $reg[1];
 }
-else {
+else
+{
     dolibarr_print_error($db);
 }
 
@@ -66,22 +71,28 @@ if ($_POST["action"] == 'add' || $_POST["modify"])
 		if ($_POST["modify"])
 		{
 			// Supprime boite box_external_rss de définition des boites
-	        $sql = "DELETE FROM ".MAIN_DB_PREFIX."boxes_def";
+	        $sql = "UPDATE ".MAIN_DB_PREFIX."boxes_def";
+			$sql.= " SET name = '".$boxlabel."'";
 	        $sql.= " WHERE file ='box_external_rss.php' AND note like '".$_POST["norss"]." %'";
-	        if (! $db->query($sql))
+
+			$resql=$db->query($sql);
+			if (! $resql)
 	        {
 				dolibarr_print_error($db,"sql=$sql");
 				exit;
 	        }
 		}
+		else
+		{
+			// Ajoute boite box_external_rss dans définition des boites
+	        $sql = "INSERT INTO ".MAIN_DB_PREFIX."boxes_def (name, file, note)";
+			$sql.= " VALUES ('".$boxlabel."','box_external_rss.php','".addslashes($_POST["norss"].' ('.$_POST[$external_rss_title]).")')";
+	        if (! $db->query($sql))
+	        {
+	            $err++;
+	        }
+		}
 		
-		// Ajoute boite box_external_rss dans définition des boites
-        $sql = "INSERT INTO ".MAIN_DB_PREFIX."boxes_def (name, file, note) VALUES ('".$boxlabel."','box_external_rss.php','".addslashes($_POST["norss"].' ('.$_POST[$external_rss_title]).")')";
-        if (! $db->query($sql))
-        {
-            $err++;
-        }
-
 		$result1=dolibarr_set_const($db, "EXTERNAL_RSS_TITLE_" . $_POST["norss"],$_POST[$external_rss_title]);
 		if ($result1) $result2=dolibarr_set_const($db, "EXTERNAL_RSS_URLRSS_" . $_POST["norss"],$_POST[$external_rss_urlrss]);
 
@@ -107,13 +118,45 @@ if ($_POST["delete"])
         $db->begin();
 
 		// Supprime boite box_external_rss de définition des boites
-        $sql = "DELETE FROM ".MAIN_DB_PREFIX."boxes_def";
+        $sql = "SELECT rowid FROM ".MAIN_DB_PREFIX."boxes_def";
         $sql.= " WHERE file ='box_external_rss.php' AND note like '".$_POST["norss"]." %'";
-        if (! $db->query($sql))
+        
+		$resql=$db->query($sql);
+		if ($resql)
         {
+			$num = $db->num_rows($resql);
+			$i=0;
+			while ($i < $num)
+			{
+				$obj=$db->fetch_object($resql);
+
+		        $sql = "DELETE FROM ".MAIN_DB_PREFIX."boxes";
+		        $sql.= " WHERE box_id = ".$obj->rowid;
+				$resql=$db->query($sql);
+
+		        $sql = "DELETE FROM ".MAIN_DB_PREFIX."boxes_def";
+		        $sql.= " WHERE rowid = ".$obj->rowid;
+				$resql=$db->query($sql);
+
+				if (! $resql)
+				{
+					$db->rollback();
+					dolibarr_print_error($db,"sql=$sql");
+					exit;
+				}
+				
+				$i++;
+			}
+
+			$db->commit();
+		}
+		else
+		{
+			$db->rollback();
 			dolibarr_print_error($db,"sql=$sql");
 			exit;
         }
+
 
 		$result1=dolibarr_del_const($db,"EXTERNAL_RSS_TITLE_" . $_POST["norss"]);
 		if ($result1) $result2=dolibarr_del_const($db,"EXTERNAL_RSS_URLRSS_" . $_POST["norss"]);
@@ -134,13 +177,15 @@ if ($_POST["delete"])
 
 
 /*
- * Affichage du formulaire de saisie
+ * Affichage page
  */
-  
+
 llxHeader();
 
 print_fiche_titre($langs->trans("ExternalRSSSetup"), $mesg, 'setup');
 
+
+// Formulaire ajout
 print '<form name="externalrssconfig" action="external_rss.php" method="post">';
 
 print '<table class="noborder" width="100%">';
@@ -150,19 +195,19 @@ print '<td>'.$langs->trans("Example").'</td>';
 print '</tr>';
 print '<tr class="impair">';
 print '<td width="100">'.$langs->trans("Title").'</td>';
-print '<td><input type="text" name="external_rss_title_'.$nbexternalrss.'" value="'.@constant("EXTERNAL_RSS_TITLE_" . $nbexternalrss).'" size="64"></td>';
+print '<td><input type="text" name="external_rss_title_'.($lastexternalrss+1).'" value="'.@constant("EXTERNAL_RSS_TITLE_" . ($lastexternalrss+1)).'" size="64"></td>';
 print '<td>April,<br>LinuxFR,<br>Lolix</td>';
 print '</tr>';
 ?>
 <tr class="pair">
   <td>URL du RSS</td>
-  <td><input type="text" name="external_rss_urlrss_<?php echo $nbexternalrss ?>" value="<?php echo @constant("EXTERNAL_RSS_URLRSS_" . $nbexternalrss) ?>" size="64"></td>
+  <td><input type="text" name="external_rss_urlrss_<?php echo ($lastexternalrss+1) ?>" value="<?php echo @constant("EXTERNAL_RSS_URLRSS_" . ($lastexternalrss+1)) ?>" size="64"></td>
   <td>http://wiki.april.org/RecentChanges?format=rss<br>http://linuxfr.org/backend/news/rss20.rss<br>http://back.fr.lolix.org/jobs.rss.php3
 </tr>
 <tr><td colspan="3" align="center">
 <input type="submit" class="button" value="<?php echo $langs->trans("Add") ?>">
 <input type="hidden" name="action" value="add">
-<input type="hidden" name="norss" value="<?php echo $nbexternalrss ?>">
+<input type="hidden" name="norss" value="<?php echo ($lastexternalrss+1) ?>">
 </td>
 </table>
 
@@ -174,70 +219,89 @@ print '</tr>';
 
 <?php
 
-for($i = 0; $i < $nbexternalrss; $i++)
+$sql ="select rowid, name, file, note from ".MAIN_DB_PREFIX."boxes_def";
+$sql.=" WHERE file = 'box_external_rss.php'";
+$sql.=" ORDER BY note";
+$resql=$db->query($sql);
+if ($resql)
 {
-	$var=true;
-	
-	$rss = fetch_rss( @constant("EXTERNAL_RSS_URLRSS_".$i) );
-	// fetch_rss initialise les objets suivant:
-	// print_r($rss->channel);
-	// print_r($rss->image);
-	// print_r($rss->items);
-	
-	print "<form name=\"externalrssconfig\" action=\"external_rss.php\" method=\"post\">";
-	
-	print "<tr class=\"liste_titre\">";
-	print "<td colspan=\"2\">Syndication du flux numéro " . ($i+1) . "</td>";
-	print "</tr>";
+	$num =$db->num_rows($resql);
+	$i=0;
 
-	$var=!$var;
-	print "<tr ".$bc[$var].">";
-	print "<td width=\"100\">".$langs->trans("Title")."</td>";
-	print "<td><input type=\"text\" class=\"flat\" name=\"external_rss_title_" . $i . "\" value=\"" . @constant("EXTERNAL_RSS_TITLE_" . $i) . "\" size=\"64\"></td>";
-	print "</tr>";
-
-	$var=!$var;
-	print "<tr ".$bc[$var].">";
-	print "<td>".$langs->trans("URL")."</td>";
-	print "<td><input type=\"text\" class=\"flat\" name=\"external_rss_urlrss_" . $i . "\" value=\"" . @constant("EXTERNAL_RSS_URLRSS_" . $i) . "\" size=\"64\"></td>";
-	print "</tr>";
-
-	$var=!$var;
-	print "<tr ".$bc[$var].">";
-	print "<td>".$langs->trans("Status")."</td>";
-	print "<td>";
-    if (! $rss->ERROR)
-    {
-		print '<font class="ok">'.$langs->trans("Online").'</div>';
-	}
-	else
+	while ($i < $num)
 	{
-		print '<font class="error">'.$langs->trans("Offline").'</div>';
-	}
-	print "</td>";
-	print "</tr>";
+		$obj = $db->fetch_object($resql);
+		
+	    eregi('^([0-9]+)',$obj->note,$reg);
+		$idrss = $reg[1];
+		//print "x".$idrss;
 
-	// Logo
-    if (! $rss->ERROR && $rss->image['url'])
-    {
+		$var=true;
+		
+		$rss = fetch_rss( @constant("EXTERNAL_RSS_URLRSS_".$idrss) );
+		// fetch_rss initialise les objets suivant:
+		// print_r($rss->channel);
+		// print_r($rss->image);
+		// print_r($rss->items);
+		
+		print "<form name=\"externalrssconfig\" action=\"external_rss.php\" method=\"post\">";
+		
+		print "<tr class=\"liste_titre\">";
+		print "<td colspan=\"2\">Syndication du flux numéro " . ($i+1) . "</td>";
+		print "</tr>";
+
 		$var=!$var;
 		print "<tr ".$bc[$var].">";
-		print "<td>".$langs->trans("Logo")."</td>";
-		print '<td>';
-		print '<img height="32" src="'.$rss->image['url'].'">';
-		print '</td>';
+		print "<td width=\"100\">".$langs->trans("Title")."</td>";
+		print "<td><input type=\"text\" class=\"flat\" name=\"external_rss_title_" . $idrss . "\" value=\"" . @constant("EXTERNAL_RSS_TITLE_" . $idrss) . "\" size=\"64\"></td>";
 		print "</tr>";
+
+		$var=!$var;
+		print "<tr ".$bc[$var].">";
+		print "<td>".$langs->trans("URL")."</td>";
+		print "<td><input type=\"text\" class=\"flat\" name=\"external_rss_urlrss_" . $idrss . "\" value=\"" . @constant("EXTERNAL_RSS_URLRSS_" . $idrss) . "\" size=\"64\"></td>";
+		print "</tr>";
+
+		$var=!$var;
+		print "<tr ".$bc[$var].">";
+		print "<td>".$langs->trans("Status")."</td>";
+		print "<td>";
+	    if (! $rss->ERROR)
+	    {
+			print '<font class="ok">'.$langs->trans("Online").'</div>';
+		}
+		else
+		{
+			print '<font class="error">'.$langs->trans("Offline").'</div>';
+		}
+		print "</td>";
+		print "</tr>";
+
+		// Logo
+	    if (! $rss->ERROR && $rss->image['url'])
+	    {
+			$var=!$var;
+			print "<tr ".$bc[$var].">";
+			print "<td>".$langs->trans("Logo")."</td>";
+			print '<td>';
+			print '<img height="32" src="'.$rss->image['url'].'">';
+			print '</td>';
+			print "</tr>";
+		}
+
+		print "<tr>";
+		print "<td colspan=\"2\" align=\"center\">";
+		print "<input type=\"submit\" class=\"button\" name=\"modify\" value=\"".$langs->trans("Modify")."\">";
+		print " &nbsp; ";
+		print "<input type=\"submit\" class=\"button\" name=\"delete\" value=\"".$langs->trans("Delete")."\">";
+		print "<input type=\"hidden\" name=\"norss\"  value=\"".$idrss."\">";
+		print "</td>";
+		print "</tr>";
+
+		print "</form>";
+		
+		$i++;
 	}
-
-	print "<tr>";
-	print "<td colspan=\"2\" align=\"center\">";
-	print "<input type=\"submit\" class=\"button\" name=\"modify\" value=\"".$langs->trans("Modify")."\"> &nbsp;";
-	print "<input type=\"submit\" class=\"button\" name=\"delete\" value=\"".$langs->trans("Delete")."\">";
-	print "<input type=\"hidden\" name=\"norss\"  value=\"$i\">";
-	print "</td>";
-	print "</tr>";
-
-	print "</form>";
 }
 ?>
 
@@ -249,5 +313,4 @@ for($i = 0; $i < $nbexternalrss; $i++)
 $db->close();
 
 llxFooter('$Date$ - $Revision$');
-
 ?>
